Add support for overriding some Micronaut module versions
This commit adds the ability to override the version of some
dependencies which are automatically added by the Gradle plugin.
For example, a `micronautCoreVersion` property in gradle.properties
file will override the default micronaut core version.

There are 2 places where the version can be overriden. Either via
the DSL, under the `micronaut` extension, e.g:

```
micronaut {
   dataVersion.set("...")
}
```

or via `gradle.properties` in which case the version is prefixed with
`micronaut`, e.g `micronautDataVersion`.

The modules for which we added version overrides are all
modules which have automatic dependencies:

  - core
  - aws
  - azure
  - data
  - gcp
  - http_netty
  - jaxrs
  - oraclecloud
  - security
  - servlet
  - validation

Note that because of https://github.com/gradle/gradle/issues/24719,
the DSL doesn't let us use the `=` operator, so one has to write
`.set("...")`. This problem is not present with the experimental
Kotlin DSL assignment operator in Gradle 8.1.

Closes #689