Add a plugin to automatically apply the Micronaut Platform catalog
This commit introduces a new _settings_ plugin, which automatically
imports the Micronaut Platform version catalog, so that the user
can declare dependencies using the `mn.xxx` notation.

This plugin needs to be applied to `settings.gradle(.kts)` file and
will determine the version of Micronaut to use either by reading
the `gradle.properties` file, in which case it will look for a couple
possible properties : `micronaut` and `micronaut.platform`, or by
reading the `micronaut` version from the user's default version
catalog.

Applying this plugin offers a new way to override the default
versions, since it is now possible to override _any_ version from
the catalog directly: unlike what is implemented in #690, it is
not limited to a known set of modules.

To override a version, the user can create a `gradle/mn.versions.toml`
file where they can override the versions found in the catalog. For
example:

```toml
[versions]
micronaut-core = "4.0.1"
micronaut-aws = "5.2.1"
```

It's worth noting that in order to be able to implement this,
we had to implement a version catalog parser, which was directly
copied from `micronaut-build`. Ideally, we should have a common
library between the 2, but there's no such thing yet.

Closes #668