Rename `mn.versions.toml` to `mn-override.versions.toml`
This makes it clearer that the TOML file is not a traditional catalog
file, but a fixup one. In addition, the error which was previously
thrown when libraries were declared in the file is now turned into a
simple warning. This will make it easier to rely on tools like
renovatebot to upgrade the versions in that file too.