Add a plugin to import the Micronaut Platform catalog (#694)
* Add support for overriding some Micronaut module versions

This commit adds the ability to override the version of some
dependencies which are automatically added by the Gradle plugin.
For example, a `micronautCoreVersion` property in gradle.properties
file will override the default micronaut core version.

There are 2 places where the version can be overriden. Either via
the DSL, under the `micronaut` extension, e.g:

```
micronaut {
   dataVersion.set("...")
}
```

or via `gradle.properties` in which case the version is prefixed with
`micronaut`, e.g `micronautDataVersion`.

The modules for which we added version overrides are all
modules which have automatic dependencies:

  - core
  - aws
  - azure
  - data
  - gcp
  - http_netty
  - jaxrs
  - oraclecloud
  - security
  - servlet
  - validation

Note that because of https://github.com/gradle/gradle/issues/24719,
the DSL doesn't let us use the `=` operator, so one has to write
`.set("...")`. This problem is not present with the experimental
Kotlin DSL assignment operator in Gradle 8.1.

Closes #689

* Add a plugin to automatically apply the Micronaut Platform catalog

This commit introduces a new _settings_ plugin, which automatically
imports the Micronaut Platform version catalog, so that the user
can declare dependencies using the `mn.xxx` notation.

This plugin needs to be applied to `settings.gradle(.kts)` file and
will determine the version of Micronaut to use either by reading
the `gradle.properties` file, in which case it will look for a couple
possible properties : `micronaut` and `micronaut.platform`, or by
reading the `micronaut` version from the user's default version
catalog.

Applying this plugin offers a new way to override the default
versions, since it is now possible to override _any_ version from
the catalog directly: unlike what is implemented in #690, it is
not limited to a known set of modules.

To override a version, the user can create a `gradle/mn.versions.toml`
file where they can override the versions found in the catalog. For
example:

```toml
[versions]
micronaut-core = "4.0.1"
micronaut-aws = "5.2.1"
```

It's worth noting that in order to be able to implement this,
we had to implement a version catalog parser, which was directly
copied from `micronaut-build`. Ideally, we should have a common
library between the 2, but there's no such thing yet.

Closes #668

* Add documentation

* Rename `mn.versions.toml` to `mn-override.versions.toml`

This makes it clearer that the TOML file is not a traditional catalog
file, but a fixup one. In addition, the error which was previously
thrown when libraries were declared in the file is now turned into a
simple warning. This will make it easier to rely on tools like
renovatebot to upgrade the versions in that file too.

* Bump Micronaut Build plugin version

(in an attempt to make CI start)