Use `.micronaut/test-resources` for the server file directory
This is done for a couple reasons. First, this aligns with what
the Maven plugin is doing. But more importantly, it will allow
a scenario where a test resources service is started in the
background, and that the services are executed from the IDE
_without_ delegating to Gradle. In that case the test resources
client will be able to find a configuration file automatically.