Initial implementation of Micronaut OpenAPI plugin (#736)
* Initial implementation of Micronaut OpenAPI plugin

This commit demonstrates an integration of the Micronaut OpenAPI generator
directly into the Gradle plugin. Because that project is still work in
progress, it uses a source dependency.

This feature allows the generation of an OpenAPI client or server directly
in the build pipeline.

For now, we only support generation of apis and models. Tests are not
supported, but it is arguable if we should handle this at the plugin
level or via something like Micronaut launch instead.

When the `io.micronaut.openapi` plugin is applied, then a new
`openapi` extension is added to the `micronaut` configuration block,
letting the user declare a definition to use. For example:

```
micronaut {
   ...
   openapi {
       client(file("src/openapi/petstore.json")) {
           apiPackageName = "com.mycompany.petstore.api"
           modelPackageName = "com.mycompany.petstore.model"
       }
   }
}
```

Then the APIs and models are generated (in 2 distinct tasks) as
part of the build process. We're using generated output directories
(e.g `build/generated/openapi/client/...`) so it means that the tasks
are executed on every invocation, but they are cacheable.

Relates to https://github.com/micronaut-projects/micronaut-openapi/issues/1037

* Manual sync of graalvm workflows

* Update dependency com.google.devtools.ksp:symbol-processing-gradle-plugin to v1.8.22-1.0.11 (#739)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update for latest OpenAPI snapshot

* Upgrade to Micronaut OpenAPI 5.0.0-M10

* Add documentation about OpenAPI plugin

* Do not use OpenAPI generator specific types as input of tasks/extensions

This will help properly isolate the execution code from the API.

* Use the worker API

While it makes the code more complex, it allows the flexibility of overriding
the default version of the generator that the plugin uses, and also makes it
possible for a user to implement their own code generators and leveraging the
plugin to perform code generation.

* Clarify documentation and fix sonar warnings

---------

Co-authored-by: Milenko Supic <supic82@gmail.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>