Add new service types to AOT optimizations by default
And you would ask: why don't you use the constant available in
`io.micronaut.aot.std.sourcegen.AbstractStaticServiceLoaderSourceGenerator`?
Good question! That's because Micronaut AOT is a _compile only_ dependency:
we have the ability to override the version of AOT which is used at runtime,
so the types that we use in the task are limited, and we cannot use any
of the AOT types at this point in time.

This means that ideally, we should have an additional AOT module which
simply exposes such constants, and use it as a runtime dependency. However,
it's a bit overkill for a single constant right now. Worse, this wouldn't
improve the situation in case the user overrides the AOT version, since
the plugin would only "see" the version it was compiled with. This means
a proper fix would be to actually read the constant dynamically, _before_
we invoke AOT. Lots of work for something which would slow down execution
and be a bit hackish.