Fix error when AOT + test resources are applied with library plugin
The Micronaut AOT plugin is expected to be applied with the application plugin.
Therefore, the test resources plugin was assuming that the application plugin
was applied, but in practice a user may apply AOT with the library plugin.
It wouldn't do anything, but it causes build errors.

This commit makes sure to wrap the relevant code in a `withPlugin` block to
avoid the error.

Fixes #815