Use an Arm compatible base image for Arm dockerBuildNative
For both Arm and X86 we are using an alpine base image.

There is no Arm variant of Alpine, so this builds an ARM native binary, then puts it in an x86 base image.

This change switches to use cgr.dev/chainguard/wolfi-base:latest when the architecture is Arm.

This image was chosen due to it's size... There may be a better default choice out there...