Add the Micronaut GraalVM dependency consistently
In previous releases of the plugin, there was an inconsistency in how the
`micronaut-graal` processor was added. In particular, the dependency was
only added if the JVM _running the build_ was GraalVM, but only for Kotlin.
For Java and Groovy, the dependency was added in any case.

In fact, the behavior for Kotlin was wrong, since there's no reason
that the toolchain used to compile the app is the same as the one
running the build. Therefore, this commit changes the behavior so that
as soon as the `io.micronaut.graalvm` plugin is applied, then the
dependency is added.

It is a slight behavioral change for users who do _not_ use GraalVM
but who would still apply the `io.micronaut.application` plugin.
Before this change, they wouldn't have a `micronaut-graalvm` annotation
processor on classpath. After this change, they will, because
the `application` plugin applies the `graalvm` plugin. It is however
significantly easier to reason about and more correct.