Fix test resources client on optimizer classpath
The AOT process should optimize applications for the production environment.
As such it doesn't make sense to have the test resources client on the
classpath of the application being scanned during the AOT optimization
process.

This bugfix has the potential to break builds of people who run
`./gradlew build` and had builds passing because the AOT optimizer was
using test resources.

Fixes #852