Workaround Kotlin 1.9.20 bug
Since Kotlin plugin 1.9.20, kapt tasks are broken if dependencies are
added to the KAPT configurations in an `afterEvaluate`. This is precisely
what we were doing. Since using `afterEvaluate` is often a code smell
(but it wasn't in our case), the proposed solution is to deprecate the
use of `additionalSourceSets`, and have an eager configuration of
sourceSets instead, as soon as the `sourceSets(...)` method is called.

It's a small breaking change, in the sense that if the user had
configured the `additionalSourceSets` and later in time decides
to remove entries from that list, then this would no longer work.