Let the user ignore automatic dependencies
In some circumstances, automatic dependencies can get in the way. This
should be rare but there was no built-in mechanism to ignore such
dependencies. It is now possible to list these using the `micronaut`
extension, for example:

```
micronaut {
    ignoredAutomaticDependencies.add("io.micronaut.data:micronaut-data-processor")
}
```

The responsibility of adding these is then shifted to the user.

Fixes https://github.com/micronaut-projects/micronaut-sql/issues/1157