Fix prepare optimizations task not being cacheable
The task which prepares optimizations makes use of a properties file
as an input (the configuration of the optimizer). Unfortunately, this
properties file, generated by the config writer task, contains a
timestamp, which breaks cacheability.

This commit fixes the problem by making sure that comments from
the writer task are removed.

Fixes #888