Create multi-layer docker files
This commit improves our docker file generation in order to use
multiple layers for dependencies. Instead of having a single libs
directory, we now have 3 layers which are copied to the `libs`
directory:

- one for the transitive dependencies which are not snapshot
- one for the snapshot dependencies
- one for the project dependencies

In addition to the layer which contains the application jar.
Moreover, the dockerfile creation now makes use of the `--link`
option for COPY, which makes it even more efficient.

Fixes #735