Add OS configuration for CRaC JDK
The Azul API has started to differentiate between linux-musl and linux-glibc.

Without us passing the option to use linux-glibc we get 2 JDKs returned, and the musl one doesn't work for building images on OSX with DockerDesktop.

This PR allows configuration of the OS when making the request to the Azul API