Remove micronaut-docs-asciidoc-extensions
This dependency is no longer available on the Gradle plugin portal
and makes the build fail.

See https://blog.gradle.org/portal-jcenter-impact for context.

I removed the dependency and tested docs generation as well as
a user application, and I noticed no difference, so I guess it
was simply redundant.