Fix GraalVM download URL
This is the Gradle's counterpart to https://github.com/micronaut-projects/micronaut-maven-plugin/pull/1219

The fix is slightly different since we already had a way to define the base path, but not the full path.
Now the plugin provides the option to completely override the GraalVM download URL, in case the derivation
from the JDK and arch properties is not good enough.