Offer the option to generate only APIs/models (#1085)
Note the calls to `task.setEnabled()` which have to be done because of
a side effect in Gradle. Basically we use:

javaMain.srcDir(clientSpec.getGenerateApis().zip(client.flatMap(DefaultOpenApiExtension::mainSrcDir), this::ifEnabled));

which will create a task dependency on the `client` generation task independently
of the result of the `enabled` flag.

Fixes #1055