Fix shadow plugin warning
The plugin warning was displayed even if the new plugin was applied,
because the new plugin also uses the old shadow id for backwards
compatibility. The fix is to check if the new shadow plugin is also
applied, but it comes with a catch: if both are applied, then there
can be a conflict, but it's likely to break with duplicate tasks
anyway, so no big deal.