Support Java Launcher for test resources service (#1181)
This commit fixes an inconsistency in the way the test resources service
was started. In practice for most projects the problem would go unnoticed,
but we have test projects which highlight the problem.

Before this change, the test resources service was systematically started
using the same Java executable as the one which was used to start Gradle
itself. Therefore, if a project was configured to run Gradle with Java 17,
but that the test resources service is compiled with Java 21, then it
woudln't start. Similarly, if any dependency of the test resources service
is not aligned with the JVM version of Gradle itself, this would fail.

This commit introduces an optional JavaLauncher property which lets the
user specify the spec of the JVM to use. Alternatively, a path to a Java
executable can also be configured in which case it takes precedence.